Better structuring of files
