+++ src/components/ReduxNetworkProvider.js
@@ -1,16 +1,19 @@
 /* @flow */
-import React, { PureComponent } from 'react';
+import { Component } from 'react';
+import { connect } from 'react-redux';
 import { AppState, NetInfo, Platform } from 'react-native';
-import NetworkContext from './NetworkContext';
-import type { HTTPMethod } from './types';
-import * as connectivityInterval from './checkConnectivityInterval';
-import checkInternetAccess from './checkInternetAccess';
-
-type State = {
-  isConnected: boolean,
-};
+import type { HTTPMethod, FluxAction } from '../types';
+import {
+  clearConnectivityCheckInterval,
+  setupConnectivityCheckInterval,
+} from '../utils/checkConnectivityInterval';
+import checkInternetAccess from '../utils/checkInternetAccess';
+import { connectionChange } from '../redux/actionCreators';
 
 type Props = {
+  dispatch: FluxAction => FluxAction,
+  isConnected: boolean,
+  actionQueue: Array<FluxAction>,
   timeout?: number,
   pingServerUrl?: string,
   shouldPing?: boolean,
@@ -39,12 +42,12 @@ function validateProps(props: Props) {
   if (typeof props.pingInBackground !== 'boolean') {
     throw new Error('you should pass a string as pingServerUrl parameter');
   }
-  if (!['HEAD', 'OPTIONS'].includes(props.httpMethod)) {
+  if (typeof !['HEAD', 'OPTIONS'].includes(props.httpMethod)) {
     throw new Error('httpMethod parameter should be either HEAD or OPTIONS');
   }
 }
 
-class NetworkProvider extends PureComponent<void, Props, State> {
+class ReduxNetworkProvider extends Component<void, Props, void> {
   static defaultProps = {
     timeout: 3000,
     pingServerUrl: 'https://www.google.com/',
@@ -58,14 +61,11 @@ class NetworkProvider extends PureComponent<void, Props, State> {
   constructor(props: Props) {
     super(props);
     validateProps(props);
-    this.state = {
-      isConnected: true,
-    };
   }
 
   async componentDidMount() {
     const { pingInterval } = this.props;
-    const handler = this.getConnectionChangeHandler();
+    const handler = this.connectionChangeHandler;
 
     NetInfo.isConnected.addEventListener('connectionChange', handler);
     // On Android the listener does not fire on startup
@@ -73,18 +73,19 @@ class NetworkProvider extends PureComponent<void, Props, State> {
       const netConnected = await NetInfo.isConnected.fetch();
       handler(netConnected);
     }
+
     if (pingInterval > 0) {
-      connectivityInterval.setup(this.intervalHandler, pingInterval);
+      setupConnectivityCheckInterval(this.intervalHandler, pingInterval);
     }
   }
 
   componentWillUnmount() {
-    const handler = this.getConnectionChangeHandler();
+    const handler = this.connectionChangeHandler;
     NetInfo.isConnected.removeEventListener('connectionChange', handler);
-    connectivityInterval.clear();
+    clearConnectivityCheckInterval();
   }
 
-  getConnectionChangeHandler() {
+  get connectionChangeHandler() {
     return this.props.shouldPing
       ? this.handleNetInfoChange
       : this.handleConnectivityChange;
@@ -112,10 +113,9 @@ class NetworkProvider extends PureComponent<void, Props, State> {
   };
 
   intervalHandler = () => {
-    if (this.state.isConnected && this.props.pingOnlyIfOffline === true) {
-      return;
+    if (!(this.props.isConnected && this.props.pingOnlyIfOffline)) {
+      this.checkInternet();
     }
-    this.checkInternet();
   };
 
   /**
@@ -123,18 +123,25 @@ class NetworkProvider extends PureComponent<void, Props, State> {
    * @param isConnected
    */
   handleConnectivityChange = (isConnected: boolean) => {
-    this.setState({
-      isConnected,
-    });
+    const { isConnected: wasConnected, actionQueue, dispatch } = this.props;
+
+    if (isConnected !== wasConnected) {
+      dispatch(connectionChange(isConnected));
+    }
+    // dispatching queued actions in order of arrival (if we have any)
+    if (!wasConnected && isConnected && actionQueue.length > 0) {
+      actionQueue.forEach((action: *) => {
+        dispatch(action);
+      });
+    }
   };
 
   render() {
-    return (
-      <NetworkContext.Provider value={this.state}>
-        {this.props.children}
-      </NetworkContext.Provider>
-    );
+    return this.props.children;
   }
 }
 
-export default NetworkProvider;
+export default connect((state: *) => ({
+  isConnected: state.network.isConnected,
+  actionQueue: state.network.actionQueue,
+}))(ReduxNetworkProvider);
