+++ src/utils/checkConnectivityInterval.js
