+++ src/utils/checkInternetConnection.js
